requirements.txt
fastapi
uvicorn
mediapipe
numpy
pillow
